Доработка программы в рамках ДЗ 7 урока
